[Android] Move training part to its own package
