+++ extension/android/executorch_android/src/main/java/org/pytorch/executorch/training/SGD.java
@@ -1,18 +1,22 @@
 /*
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- * All rights reserved.
  *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree.
+ *  * Copyright (c) Meta Platforms, Inc. and affiliates.
+ *  * All rights reserved.
+ *  *
+ *  * This source code is licensed under the BSD-style license found in the
+ *  * LICENSE file in the root directory of this source tree.
+ *
+ *
  */
 
-package org.pytorch.executorch;
+package org.pytorch.executorch.training;
 
 import com.facebook.jni.HybridData;
 import com.facebook.jni.annotations.DoNotStrip;
 import com.facebook.soloader.nativeloader.NativeLoader;
 import com.facebook.soloader.nativeloader.SystemDelegate;
 import java.util.Map;
+import org.pytorch.executorch.Tensor;
 import org.pytorch.executorch.annotations.Experimental;
 
 /**
@@ -62,7 +66,7 @@ public class SGD {
    * @param dampening The dampening value
    * @param weightDecay The weight decay value
    * @param nesterov Whether to use Nesterov momentum
-   * @return new {@link org.pytorch.executorch.SGD} object
+   * @return new {@link SGD} object
    */
   public static SGD create(
       Map<String, Tensor> namedParameters,
@@ -79,7 +83,7 @@ public class SGD {
    *
    * @param namedParameters Map of parameter names to tensors to be optimized
    * @param learningRate The learning rate for the optimizer
-   * @return new {@link org.pytorch.executorch.SGD} object
+   * @return new {@link SGD} object
    */
   public static SGD create(Map<String, Tensor> namedParameters, double learningRate) {
     return create(namedParameters, learningRate, 0.0, 0.0, 0.0, false);
